+++ CHANGELOG.md
@@ -9,14 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- All CRDs are now available as both `v1.CustomResourcDefinition` and `v1beta1.CustomResourceDefinition` through 
+  `crd.LoadV1` and `crd.LoadV1Beta1`. Type-specific `New*CRD()` functions are unchanged.
 
 ## [0.3.3] - 2020-04-28
 
 ### Added
 
 - Generate Cluster API CRDs from upstream module.
-
-
 
 ## [0.3.2] - 2020-04-27