+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Add Azure Tools CRDs
+## [0.3.5] - 2020-05-06
+
+### Added
+
+- Add Azure Tools CRDs.
 
 ## [0.3.4] - 2020-04-30
 
@@ -159,7 +163,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - First release.
 
 
-[Unreleased]: https://github.com/giantswarm/apiextensions/compare/v0.3.4...HEAD
+[Unreleased]: https://github.com/giantswarm/apiextensions/compare/v0.3.5...HEAD
+[0.3.5]: https://github.com/giantswarm/apiextensions/compare/v0.3.4...v0.3.5
 [0.3.4]: https://github.com/giantswarm/apiextensions/compare/v0.3.3...v0.3.4
 [0.3.3]: https://github.com/giantswarm/apiextensions/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/giantswarm/apiextensions/compare/v0.3.1...v0.3.2